@@ -16,23 +16,21 @@
         }
     ],
     "minimum-stability": "stable",
-
+    
     "require": {
-        "php": "*",
+        "php": "^7.2.5",
+        "ext-ctype": "*",
         "ext-iconv": "*",
         "api-platform/api-pack": "^1.2",
-        "chgk/chgkdb-parser": "v1.0.9",
-        "symfony/console": "^4.0",
-        "symfony/flex": "^1.0",
-        "symfony/framework-bundle": "^4.0",
-        "symfony/lts": "^4@dev",
-        "symfony/monolog-bundle": "^3.0@dev",
-        "symfony/security-bundle": "^4.0",
-        "symfony/web-server-bundle": "^4.0",
-        "symfony/yaml": "^4.0"
+        "chgk/chgkdb-parser": "v1.*",
+        "symfony/console": "5.0.*",
+        "symfony/dotenv": "5.0.*",
+        "symfony/flex": "^1.3.1",
+        "symfony/framework-bundle": "5.0.*",
+        "symfony/monolog-bundle": "^3.5",
+        "symfony/yaml": "5.0.*"
     },
     "require-dev": {
-        "symfony/dotenv": "^4.0"
     },
     "config": {
         "preferred-install": {
@@ -51,7 +49,10 @@
         }
     },
     "replace": {
+        "paragonie/random_compat": "2.*",
+        "symfony/polyfill-ctype": "*",
         "symfony/polyfill-iconv": "*",
+        "symfony/polyfill-php72": "*",
         "symfony/polyfill-php71": "*",
         "symfony/polyfill-php70": "*",
         "symfony/polyfill-php56": "*"
@@ -59,7 +60,7 @@
     "scripts": {
         "auto-scripts": {
             "cache:clear": "symfony-cmd",
-            "assets:install --symlink --relative %PUBLIC_DIR%": "symfony-cmd"
+            "assets:install %PUBLIC_DIR%": "symfony-cmd"
         },
         "post-install-cmd": [
             "@auto-scripts"
@@ -73,8 +74,8 @@
     },
     "extra": {
         "symfony": {
-            "id": "01CC98WQPXM1Y8E6GWHJTEP60D",
-            "allow-contrib": false
+            "allow-contrib": false,
+            "require": "5.0.*"
         }
     }
 }
